refactor(settings): split the monolith into shell + SettingsStore + 11 section components
Stage 1 of the model-settings unification (docs/research/2026-07-02-unify-model-settings.md).
Mechanical, behavior-identical — no copy, control, or payload change:
- settings.component.ts becomes a 479-line shell (sidebar search/nav + Save + @switch);
  selector and router import path unchanged
- NEW settings.store.ts (provided on the shell, route lifetime): the single typed FormGroup
  (31 controls, same names/defaults), all cross-section signals/computeds (providerIsCloud,
  gateway state, download progress), load()/save()/keys/consents/downloads, event-stream
  subscriptions with DestroyRef cleanup; writable signals privatized (_x + asReadonly)
- NEW sections/settings-*-section.component.ts × 11: verbatim template+style slices,
  :host{display:contents}, readonly aliases onto the store
- kills the 15.51/16 kB style-budget cliff: largest child is now 4.3 kB compiled

Verified: ng lint + ng build green; adversarial verifier PASS — save_config payload
byte-equal pre/post split (37 keys, identical set AND order) on independent fixtures incl.
hidden-control and key-control round-trips; route leave/re-enter = clean unlisten + fresh
load (no subscription pile-up); rapid section switching ×2 clean (no NG0600/ɵcmp); in-flight
download state survives section destruction; 8/11 sections pixel-identical, rest subpixel
antialiasing; stage-0 copy fixes survived verbatim; scripts/ci.sh green on the integrated
tree.
